create GraphQLError.h
move schema_exception to its own file and namespace, it's a base class
that does not know about response::Value.

GraphQLService.h still defines its own subclass that imlements
getErrors() returning response::Value, as well as redeclaring the
moved types (field_path, path_segment) to keep existing code working.

what() was changed to use the structuredErrors rather than the
response::Value, it's more generic and cheaper.
