+++ src/GraphQLService.cpp
@@ -93,58 +93,14 @@ response::Value buildErrorValues(const std::vector<schema_error>& structuredErro
 }
 
 schema_exception::schema_exception(std::vector<schema_error>&& structuredErrors)
-	: _structuredErrors(std::move(structuredErrors))
-	, _errors(buildErrorValues(_structuredErrors))
+	: graphql::error::schema_exception(std::move(structuredErrors))
+	, _errors(buildErrorValues(std::as_const(*this).getStructuredErrors()))
 {
 }
 
 schema_exception::schema_exception(std::vector<std::string>&& messages)
-	: schema_exception(convertMessages(std::move(messages)))
+	: graphql::service::schema_exception(convertMessages(std::move(messages)))
 {
-}
-
-std::vector<schema_error> schema_exception::convertMessages(
-	std::vector<std::string>&& messages) noexcept
-{
-	std::vector<schema_error> errors(messages.size());
-
-	std::transform(messages.begin(),
-		messages.end(),
-		errors.begin(),
-		[](std::string& message) noexcept {
-			return schema_error { std::move(message) };
-		});
-
-	return errors;
-}
-
-const char* schema_exception::what() const noexcept
-{
-	const char* message = nullptr;
-
-	if (_errors.size() > 0)
-	{
-		auto itr = _errors[0].find("message");
-
-		if (itr != _errors[0].end() && itr->second.type() == response::Type::String)
-		{
-			message = itr->second.get<response::StringType>().c_str();
-		}
-	}
-
-	return (message == nullptr) ? "Unknown schema error" : message;
-}
-
-const std::vector<schema_error>& schema_exception::getStructuredErrors() const noexcept
-{
-	return _structuredErrors;
-}
-
-std::vector<schema_error> schema_exception::getStructuredErrors() noexcept
-{
-	auto structuredErrors = std::move(_structuredErrors);
-
-	return structuredErrors;
 }
 
 const response::Value& schema_exception::getErrors() const noexcept
